requirements.txt
smbus2
fastapi
uvicorn
pydantic
prometheus_client
paho-mqtt
python-dateutil
